# Add option to use `ripgrep` for crawling the list of files