--- a/lib/load-paths-handler.js
+++ b/lib/load-paths-handler.js
@@ -21,7 +21,7 @@ class PathLoader {
     this.paths = []
     this.inodes = new Set()
     this.repo = null
-    if (ignoreVcsIgnores) {
+    if (ignoreVcsIgnores && !this.useRipGrep) {
       const repo = GitRepository.open(this.rootPath, {refreshOnWindowFocus: false})
       if ((repo && repo.relativize(path.join(this.rootPath, 'test'))) === 'test') {
         this.repo = repo