--- a/lib/load-paths-handler.js
+++ b/lib/load-paths-handler.js
@@ -5,16 +5,19 @@ const fs = require('fs')
 const path = require('path')
 const {GitRepository} = require('atom')
 const {Minimatch} = require('minimatch')
+const childProcess = require('child_process')
+const { rgPath } = require('vscode-ripgrep')
 
 const PathsChunkSize = 100
 
 const emittedPaths = new Set()
 
 class PathLoader {
-  constructor (rootPath, ignoreVcsIgnores, traverseSymlinkDirectories, ignoredNames) {
+  constructor (rootPath, ignoreVcsIgnores, traverseSymlinkDirectories, ignoredNames, useRipGrep) {
     this.rootPath = rootPath
     this.traverseSymlinkDirectories = traverseSymlinkDirectories
     this.ignoredNames = ignoredNames
+    this.useRipGrep = useRipGrep
     this.paths = []
     this.inodes = new Set()
     this.repo = null
@@ -27,10 +30,50 @@ class PathLoader {
   }
 
   load (done) {
+    if (this.useRipGrep) {
+      this.loadFromRipGrep().then(done)
+
+      return
+    }
+
     this.loadPath(this.rootPath, true, () => {
       this.flushPaths()
       if (this.repo != null) this.repo.destroy()
       done()
+    })
+  }
+
+  async loadFromRipGrep () {
+    return new Promise((resolve) => {
+      const args = ['--files', '--hidden', '--sort', 'path']
+
+      if (this.ignoreVcsIgnores) {
+        args.push('--no-ignore')
+      }
+
+      if (this.traverseSymlinkDirectories) {
+        args.push('--follow')
+      }
+
+      for (let ignoredName of this.ignoredNames) {
+        args.push('-g', '!' + ignoredName.pattern)
+      }
+
+      let output = ''
+      const result = childProcess.spawn(rgPath, args, {cwd: this.rootPath})
+
+      result.stdout.on('data', chunk => {
+        const files = (output + chunk).split('\n')
+        output = files.pop()
+
+        for (const file of files) {
+          this.pathLoaded(path.join(this.rootPath, file))
+        }
+      })
+      result.on('close', () => {
+        this.flushPaths()
+        resolve()
+      })
     })
   }
 
@@ -54,7 +97,7 @@ class PathLoader {
     if (this.paths.length === PathsChunkSize) {
       this.flushPaths()
     }
-    done()
+    done && done()
   }
 
   flushPaths () {
@@ -114,7 +157,7 @@ class PathLoader {
   }
 }
 
-module.exports = function (rootPaths, followSymlinks, ignoreVcsIgnores, ignores = []) {
+module.exports = function (rootPaths, followSymlinks, ignoreVcsIgnores, ignores, useRipGrep) {
   const ignoredNames = []
   for (let ignore of ignores) {
     if (ignore) {
@@ -133,7 +176,8 @@ module.exports = function (rootPaths, followSymlinks, ignoreVcsIgnores, ignores 
         rootPath,
         ignoreVcsIgnores,
         followSymlinks,
-        ignoredNames
+        ignoredNames,
+        useRipGrep
       ).load(next)
     ,
     this.async()