--- a/lib/load-paths-handler.js
+++ b/lib/load-paths-handler.js
@@ -2,6 +2,7 @@
 
 const async = require('async')
 const fs = require('fs')
+const os = require('os')
 const path = require('path')
 const {GitRepository} = require('atom')
 const {Minimatch} = require('minimatch')
@@ -9,6 +10,10 @@ const childProcess = require('child_process')
 const { rgPath } = require('vscode-ripgrep')
 
 const PathsChunkSize = 100
+
+// Define the maximum number of concurrent crawling processes based on the number of CPUs
+// with a maximum value of 8 and minimum of 1.
+const MaxConcurrentCrawls = Math.min(Math.max(os.cpus().length - 1, 8), 1)
 
 const emittedPaths = new Set()
 
@@ -169,8 +174,9 @@ module.exports = function (rootPaths, followSymlinks, ignoreVcsIgnores, ignores,
     }
   }
 
-  async.each(
+  async.eachLimit(
     rootPaths,
+    MaxConcurrentCrawls,
     (rootPath, next) =>
       new PathLoader(
         rootPath,