--- a/lib/load-paths-handler.js
+++ b/lib/load-paths-handler.js
@@ -11,6 +11,9 @@ const { rgPath } = require('vscode-ripgrep')
 
 const PathsChunkSize = 100
 
+// Use the unpacked path if the ripgrep binary is in asar archive.
+const realRgPath = rgPath.replace(/\bapp\.asar\b/, 'app.asar.unpacked')
+
 // Define the maximum number of concurrent crawling processes based on the number of CPUs
 // with a maximum value of 8 and minimum of 1.
 const MaxConcurrentCrawls = Math.min(Math.max(os.cpus().length - 1, 8), 1)
@@ -65,7 +68,7 @@ class PathLoader {
       }
 
       let output = ''
-      const result = childProcess.spawn(rgPath, args, {cwd: this.rootPath})
+      const result = childProcess.spawn(realRgPath, args, {cwd: this.rootPath})
 
       result.stdout.on('data', chunk => {
         const files = (output + chunk).split('\n')