--- a/lib/path-loader.js
+++ b/lib/path-loader.js
@@ -10,13 +10,16 @@ module.exports = {
     ignoredNames = ignoredNames.concat(atom.config.get('core.ignoredNames') || [])
     const ignoreVcsIgnores = atom.config.get('core.excludeVcsIgnoredPaths')
     const projectPaths = atom.project.getPaths().map((path) => fs.realpathSync(path))
+    const useRipGrep = atom.config.get('fuzzy-finder.useRipGrep')
 
     const task = Task.once(
       taskPath,
       projectPaths,
       followSymlinks,
       ignoreVcsIgnores,
-      ignoredNames, () => callback(results)
+      ignoredNames,
+      useRipGrep,
+      () => callback(results)
     )
 
     task.on('load-paths:paths-found',