--- a/package.json
+++ b/package.json
@@ -15,6 +15,7 @@
     "minimatch": "~3.0.3",
     "temp": "~0.8.1",
     "underscore-plus": "^1.0.0",
+    "vscode-ripgrep": "^1.2.5",
     "wrench": "^1.5"
   },
   "devDependencies": {
@@ -74,6 +75,11 @@
       "type": "boolean",
       "default": true,
       "description": "Use an alternative scoring approach which prefers run of consecutive characters, acronyms and start of words. (Experimental)"
+    },
+    "useRipGrep": {
+      "type": "boolean",
+      "default": false,
+      "description": "Use the experimental `ripgrep` crawler. This will speed up substantially the indexing process on large projects."
     },
     "prefillFromSelection": {
       "type": "boolean",