--- a/package.json
+++ b/package.json
@@ -19,6 +19,7 @@
     "wrench": "^1.5"
   },
   "devDependencies": {
+    "sinon": "1.17.4",
     "standard": "^10.0.3"
   },
   "standard": {