## Summary

This PR contains some minimal changes to demonstrate the usage of `ripgrep` on the fuzzy finder, so we can discuss the tradeoffs and steps needed to be able to ship this.

This work comes as a follow-up from https://github.com/atom/fuzzy-finder/pull/367 after the suggestions of @smashwilson .

## Benefits

Using `ripgrep` speeds up drastically the time to crawl medium and large repositories (we're taking about **up to 14X faster times**):

| Type | Num Files | Time (master)  | Time (This PR) | Improvements  |
|---|---|---|---|---|
|  Small | 2K  | 1.3s  | 0.9s | **30%** less time (or **1.4X** faster)  🎉 |
|  Medium |  30K | 6.2s  | 1.2s  | **80%** less time (or **5X** faster) 🎉 |
|  Large | 270K  |  57.5s  |  4.1s  |  **93%** less time (or **14X** faster) 🎉 |

(the measurement has been done the same way as in https://github.com/atom/fuzzy-finder/pull/366).

## Possible Drawbacks

The crawling behaviour with `ripgrep` is slightly different than the one currently implemented. I don't think that the changes are important enough (or even bad), but it's important to list them:

1. `ripgrep` also returns symlink destination files (e.g if there's a symlink `./foo.js` which points to `./bar.js`, with `ripgrep` `foo.js` can also be opened. I think this is an improvement.
2. The ordering of the returned results is slightly different: right now results are returned in a DFS ordering (then alphabetically), while `ripgrep` returns all results alphabetically ordered.
3. Currently, if Atom gets opened from a non-git folder which has git repositories in some sub-folders, the `.gitignore` files from the sub-folders are ignored. With `ripgrep` they are taken into account. I think that this is an improvement.

Regarding 2., this change is quite noticeable, since the fuzzy finder currently displays the first 10 files returned by the crawler when it gets opened. This means that the first shown files will be slightly different than before (IMO we should change the default list of files that are shown when opening the fuzzy finder to show e.g recently opened files or currently opened files).

## Alternate Designs

Other potential designs were presented [here](https://github.com/atom/fuzzy-finder/pull/367), and seems like `ripgrep` is the best tool to use at this moment.

Regarding the current implementation, I've decided to use [`vscode-ripgrep`](https://github.com/roblourens/vscode-ripgrep), which is just a package that handles the download of the `ripgrep` binaries for several platforms. This has saved me a ton of time, but I'm not familiar enough with the process to install bundled extensions in Atom to know if this  package is suitable for the job.

## Next steps

- [x] Write some automated tests to verify the `ripgrep` crawling.
- [x] Manually check that `ripgrep` works well on different scenarios.
  - [x] Different types of file system (FAT, NTFS, ...).
  - [x] Different OSes (Windows).
  - [x] Check that the `ripgrep` binary can be used correctly from a production build (e.g does the asar packaging cause any issue?).
    - [x] OSX
    - [x] Windows
- [x] Check that the behaviour is correct.
  - [x] Files/folders with utf8 names.
  - [x] List of returned files is the same in a few different repos. 

## Applicable Issues

 https://github.com/atom/fuzzy-finder/issues/271