@rafeca 2019-03-12
I've created https://github.com/atom/fuzzy-finder/pull/369 as a follow-up PR based on this discussion, closing this PR