@rafeca 2019-03-13
**Small update:** I just realized that when using `ripgrep` we don't need to initialize the `Git` native code (see [commit](https://github.com/atom/fuzzy-finder/pull/369/commits/937a3bea6ca6177d57e5ebc37efa3075b0ad9190)) 😃

This is quite slow (specially for large repos) so it gives us even greater performance improvements (we can reduce the times on large repos by another 50% 🚀). I'm gonna update the results table with the new data