@rafeca 2019-03-13
I think it's good to share a couple of screencasts to compare the previous experience on very large repos (`gecko-dev`) with the one after this PR:

What happens **now in master** after opening Atom (or focusing Atom back):

![master](https://user-images.githubusercontent.com/408035/54277439-292e0600-4590-11e9-8c86-8b231c41d963.gif)

The same flow **with this PR**:

![ripgrep](https://user-images.githubusercontent.com/408035/54277525-66929380-4590-11e9-9e19-a490fe7003b1.gif)