@jasonrudolph 2019-03-13
> Currently, if Atom gets opened from a non-git folder which has git repositories in some sub-folders, the .gitignore files from the sub-folders are ignored. With ripgrep they are taken into account. I think that this is an improvement.

I consider this a **huge** improvement. 😍

> I think it's good to share a couple of screencasts to compare the previous experience on very large repos (gecko-dev) with the one after this PR...

Wow! 🏎💨

> Next steps...

Thanks for listing these out. This seems promising and very much worth pursuing. :zap: