@rafeca 2019-03-14
The testing so far seems to be going well:

* I've modified the test suite to run all the fuzzy finder tests both with and without `ripgrep`, this way we have good coverage for `ripgrep`.
* The functionality works fine on Windows and on different file systems.
* I've been able to test a prod build on OSX and in fact I had to do a small fix to support asar archives.
* I haven't been able to build Atom on Windows yet... This is the last bit of remaining work.

A couple of drawbacks that I've discovered today:

* The `ripgrep` binary is quite big (~5MB for OSX, a little bit less for Windows). Once it gets `gzipped` its size gets reduced to ~1.8MB, which means that the install files of Atom will grow by ~1.25% (based on the current size of ~144MB).
* Currently, `vscode-ripgrep` downloads the needed binary from GitHub during installation ([code](https://github.com/roblourens/vscode-ripgrep/blob/master/lib/download.js#L33)). This causes failures on CI systems due to rate limits. I've fixed that for now by putting setting a GitHub token of mine as an env var on the travis config for `fuzzy-finder`. We should either do the same in `atom/atom` or stop using `vscode-ripgrep` and build (and maintain) our own package.

@jasonrudolph any thoughts about these drawbacks?