@rafeca 2019-03-14
Awesome! I agree it's a worthwhile tradeoff 😃

I've been able to verify that Windows builds are packaged correctly with `ripgrep`, which was the last item from my list of things to consider... So this PR is ready for review 😄