@rafeca 2019-03-14
Oh there's something else, this feature is behind a config param which is off by default. Since it's a fuzzy-finder config param, the setting is quite hidden and really hard to discover (you have to go to settings -> packages -> search fuzzy finder and click on its settings.

This is how this setting looks like:

<img width="771" alt="Screen Shot 2019-03-14 at 21 54 07" src="https://user-images.githubusercontent.com/408035/54390888-c6c62a00-46a3-11e9-95f3-eea79f2c5a08.png">

I have mixed feelings about this:

* The benefit is that by having the setting we can find early issues from people that opt in before we turn it on by default.
* The problem is that the setting is so hidden that I don't expect many people to turn it on, so we're delaying the impact of it until we make it on by default (which takes at least one month) without getting much more testing in exchange.

Couple of things that we could do:

1. Turn it on by default (and remove the "experimental" wording from the setting). If under some situations something goes wrong we can tell people to turn it off as a workaround.
2. Leave it as it is: one month is not that much time and even a little bit of extra testing is appreciated.
3. Leave it as off but implement a popup that prompts users to turn it on if we detect that they open a medium or large repository. This is my favorite one, but if this takes much effort to implement then I don't think it's worth it.