@jasonrudolph 2019-03-18
> Oh there's something else, this feature is behind a config param which is off by default. Since it's a fuzzy-finder config param, the setting is quite hidden and really hard to discover...
>
> Couple of things that we could do:
>
> 1. ...
> 2. ...
> 3. Leave it as off but implement a popup that prompts users to turn it on if we detect that they open a medium or large repository. This is my favorite one, but if this takes much effort to implement then I don't think it's worth it.

I love this idea! It doesn't need to be part of this PR, but I think it's worth trying in a follow-up PR. I'd be happy to help. :smile:

For what it's worth, I _think_ it will be fairly straightforward to implement using the [notification API](https://atom.io/docs/api/v1.35.1/NotificationManager).